refactor(contract): move GuestLibc below the crates that need it
Two crates that cannot see each other both need to name a guest's C
library. `mvm-build` detects it from an unpacked tree, and `mvm-fs` has
to key the SDK sidecar cache on it so a guest is offered the variant it
can actually load. Those are siblings — `mvm-fs` depends only on
`mvm-contract` and `mvm-http` — so the type has to sit underneath both.

Move the enum to `mvm-contract` and re-export it from `mvm-build`, whose
callers keep their paths. Detection stays where it was: it reads a
filesystem, which a `no_std` crate cannot.

The type gains a `Display` impl and a test asserting it agrees with the
serde form, because the cache is about to key a directory segment on
this: a divergence between the two would put the artifact somewhere the
resolver does not look, and nothing else would notice.

Refs #2969.
